refactor: simplify the fluent-editor.ts file, retaining only the logic related to module registration
