+++ packages/fluent-editor/src/ui/icons.config.ts
